+++ test/src/main/scala/japgolly/scalajs/react/test/package.scala
@@ -8,12 +8,12 @@ package object test {
   final type ComponentClass = ReactComponentC_
   final type ComponentM = ReactComponentM_[TopNode]
 
-  sealed trait ComponentOrNode extends Object
-  @inline final implicit def autoComponentOrNodeN(n: dom.Node): ComponentOrNode =
-    n.asInstanceOf[ComponentOrNode]
-  @inline final implicit def autoComponentOrNodeU(c: ReactComponentU_): ComponentOrNode =
-    c.asInstanceOf[ComponentOrNode]
-  @inline final implicit def autoComponentOrNodeM[N <: TopNode](c: ReactComponentM_[N]): ComponentOrNode =
+  sealed trait ReactOrDomNode extends Object
+  @inline final implicit def autoReactOrDomNodeN(n: dom.Node): ReactOrDomNode =
+    n.asInstanceOf[ReactOrDomNode]
+  @inline final implicit def autoReactOrDomNodeU(c: ReactElement): ReactOrDomNode =
+    c.asInstanceOf[ReactOrDomNode]
+  @inline final implicit def autoReactOrDomNodeM[N <: TopNode](c: ReactComponentM_[N]): ReactOrDomNode =
     c.getDOMNode()
 
   @inline final implicit def RTUSChangeEventData  (d: ChangeEventData  ): Object = d.toJs
@@ -21,9 +21,9 @@ package object test {
   @inline final implicit def RTUSMouseEventData   (d: MouseEventData   ): Object = d.toJs
 
   @inline final implicit def autoUnboxRefsInTests[T <: TopNode](r: UndefOr[ReactComponentM_[T]]) = r.get
-  @inline final implicit def autoUnboxRefsInTestsC[T <: TopNode](r: UndefOr[ReactComponentM_[T]]): ComponentOrNode = r.get
+  @inline final implicit def autoUnboxRefsInTestsC[T <: TopNode](r: UndefOr[ReactComponentM_[T]]): ReactOrDomNode = r.get
 
 //  implicit final class RTUSimulateExt(val u: Simulate) extends AnyVal {
-//    def change(t: ComponentOrNode, newValue: String) = u.change(t, ChangeEventData(value = newValue))
+//    def change(t: ReactOrDomNode, newValue: String) = u.change(t, ChangeEventData(value = newValue))
 //  }
 }